Make eideticker an installable package
Currently we "install" the eideticker server as a script, you must manually
create a "captures" directory for it to dump its data into. For our initial
use-case, it looks like it will be easier to control the capturing directly
from the test harness, so I'll probably be removing the server for now
shortly.